Updated "Place" to "Context", removed "lists"
